web_personal/requirements.txt
asgiref==3.5.0
astroid==2.11.2
autopep8==1.6.0
dill==0.3.4
Django==4.0.3
django-ckeditor==6.2.0
django-js-asset==2.0.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pylint==2.13.4
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
wrapt==1.14.0
